Alumni myself (dwijnand)
It's been an honour and a privilege to collaborate on Cargo and Rustup,
but my interests and attention have been elsewhere recently so I feel
I'd be doing my teammates a disservice by continuing to occupy a team
slot.  I continue to wish the entire Rust project success and it's quite
possible that I might come back in the future.  Thank you, Dale.