docsrs-ops: remove team
The team was originally created as a way to give people with access to
the docs.rs machine access to our Grafana instance. Now that a proper
docs-rs team was created there is no need for this anymore.

Grafana is already switched to grant access to the docs-rs team instead
of the docsrs-ops one.