alumni: avoid including "partial alumni" in the alumni team
There are some members of the Rust project that retired from one team
but are still active in other teams. Before this commit they were listed
in the "alumni" page, as they were marked as alumni of the team they
retired of, but that didn't feel right as they haven't fully left the
project.

This commit excludes members that are active in any team from being
listed in the alumni team, even if they are alumni of one team.