Add a command to show information about a person
It shows their name, GitHub username, email (if present), and all of the
teams (including WGs) that they're a member of.

It's like `dump-team`, but for people.

The output looks like this:

    -- Camelid --

    github: @camelid

    teams:
      - icebreakers-cleanup-crew
      - icebreakers-llvm
      - wg-prioritization
      - wg-triage

or, for someone on more teams and with an email:

    -- Pietro Albini --

    github: @pietroalbini
    email:  pietro@pietroalbini.org

    teams:
      - all
      - core
      - crates-io
      - docs-rs
      - infra
      - infra-admins
      - inside-rust-reviewers
      - leads
      - project-foundation
      - release
      - website
      - wg-rustfix
      - wg-security-response

or, for someone on no teams (rustbot's not actually in the teams data, I
created an entry for it just for this purpose):

    -- rustbot --

    github: @rustbot

    teams:
      (none)