mod team resignation
The entire moderation team resigns, effective immediately. This resignation is
done in protest of the Core Team placing themselves unaccountable to anyone but
themselves.

As a result of such structural unaccountability, we have been unable to enforce
the Rust Code of Conduct to the standards the community expects of us and
to the standards we hold ourselves to. To leave under these circumstances
deeply pains us, and we apologize to all of those that we have let down. In
recognition that we are out of options from the perspective of Rust Governance,
we feel as though we have no course remaining to us but to step down and make
this statement.

In so doing, we would offer a few suggestions to the community writ large:

* We suggest that Rust Team Members come to a consensus on a process for
oversight over the Core Team. Currently, they are answerable only to
themselves, which is a property unique to them in contrast to all other Rust
teams.
* In the interest of not perpetuating unaccountability, we recommend that the
replacement for the Mod Team be made by Rust Team Members not on the Core Team.
* We suggest that the future Mod Team, with advice from Rust Team Members,
proactively decide how best to handle and discover unhealthy conflict among
Rust Team Members. We suggest that the Mod Team work with the Foundation in
obtaining resources for professional mediation.
* Additionally, while not related to this issue, based on our experience in
moderation over the years, we suggest that the future Mod Team take special
care to keep the team of a healthy size and diversity, to the extent possible.
It is a thankless task, and we did not do our best to recruit new members.

In this message, we have avoided airing specific grievances beyond
unaccountability. We've chosen to maintain discretion and confidentiality. We
recommend that the broader Rust community and the future Mod Team exercise
extreme skepticism of any statements by the Core Team (or members thereof)
claiming to illuminate the situation.

We are open to being contacted by Rust Team Members for advice or
clarification.

Sincerely,
The Rust Moderation Team (Andre, Andrew and Matthieu)