Introduce v2 permissions API
The v1 permissions API can be misleading in that the order of the `github_users` does not match the order of the `github_ids` (see https://rust-lang.zulipchat.com/#narrow/stream/318791-t-crates-io/topic/team.20repo.20integration/near/412320079).

This change introduces a v2 permissions API which uses a list of objects instead of an object with multiple single-field lists. This allows us to easily show both GitHub login and ID of a person in the logs.