remove branch protections from regex repo
Commit 55158aeb89463e1fc2d8e5b5321810cb70757fc3 introduced automation
for the regex repo but also enabled branch protection for `master`. The
regex repo has never used branch protection and it is low traffic enough
to be annoying.

While most changes go in through a PR, I do for example push directly to
master when cutting a release (which is done through my own automation
locally).