Join and lead the Rust 2024 edition project
In the lang design meeting on 2024-04-10 (with Eric Huss and Mara
present), there formed a consensus that I would take on responsibility
for the project management of getting the Rust 2024 edition released.

We understand that bstrie is stepping away from this effort, and we
thank him for all the good work he's done to this point.

Correspondingly, in this commit, we add me (TC) to the Rust 2024
edition project as a member and a lead, and we add bstrie to our list
of alumni.