Validate homepage URLs on GitHub
Repositories can be configured with a homepage, which is prominently
featured on GitHub as a link. For repositories under the `rust-lang`
organization, we want to make sure that those links only point to
domains that are explicitly allowed. Ideally, only domains owned and
operated by the Rust project itself will be whitelisted.

The risk with other domains is that they might expire silently and get
taken over by malicious actors, who can then host phishing campaigns or
malware on sites "advertised" by the Rust project.