propose goals team
charter and other documentation available on
the [rust-project-goals repository][r]. As of
this writing:

The Rust goals program is administered by the Goals team.
This document serves as the team charter.

Our mission is to focus the Rust programming language efforts by running and administering an inclusive and dynamic goals program. We work with the project teams to identify the highest priority items and to make sure that the teams are budgeting adequate time and resources to ensure those items are successful. For new contributors who have an idea they'd like to pursue, we work to provide a welcoming "front door" to Rust, connecting their proposal to the maintainers whose support will be needed to make it reality. For existing maintainers, we help them to document the work they are doing and to find new contributors.

Team members perform some subset of the following roles:

* Attending short sync meetings.
* When preparing a new goal slate:
    * Advertising the goal program to teams and soliciting participation
    * Reviewing incoming goal proposals for quality and accuracy
    * Seeking feedback on behalf of outsiders' goals
    * Authoring the RFC and hounding team leads to check their boxes
    * Deciding which goals to propose as flagship goals.
* During the year:
    * Authoring round-up blog posts highlighting progress
    * Updating and maintaining the web-site
    * Checking in with the goal points of contact that are not reporting progress to see if they need help

The team lead is the [owner](../about/owners.md) of the program,
meaning that they take ultimately responsible for ensuring the goals program moves forward smoothly.
They perform and and all of the member functions as needed, delegating where possible.
In the event of conflicts (e.g., which goals to propose as flagship goals in the RFC), team lead makes the final decision.

[r]: https://github.com/rust-lang/rust-project-goals/blob/3ed26fd345698de857bdb67cafdd698c863a0c65/src/admin/team.md