Adopt rust-by-example as a subteam of lang-docs
We're working to move things out from under the launching-pad team to
more appropriate places.  It makes the most sense for the
rust-by-example working group to be adopted as a subteam of lang-docs,
so let's do that.

Note that, in doing this, we'll make it a proper subteam rather than a
working group.