Adopt rustlings team as a subteam of lang-docs
We're doing some minor reorganization.  As part of that, we're moving
the book team and rust-by-example to be subteams of lang-docs.  Given
that, it makes sense for lang-docs to also adopt rustlings as a
subteam, so let's do that.

Note that, unlike the other two, rustlings had been notated as a
marker team rather than a subteam of launching-pad.  This change will
make it into a normal subteam, which seems more appropriate.