Adopt rustlings team as a subteam of lang-docs (#1638)
* Adopt rustlings team as a subteam of lang-docs

We're doing some minor reorganization.  As part of that, we're moving
the book team and rust-by-example to be subteams of lang-docs.  Given
that, it makes sense for lang-docs to also adopt rustlings as a
subteam, so let's do that.

Note that, unlike the other two, rustlings had been notated as a
marker team rather than a subteam of launching-pad.  This change will
make it into a normal subteam, which seems more appropriate.

* Fill in email/Zulip info for mo8it

In the PR, mo8it provided these details to be included.

* Make mo8it a lead of rustlings subteam

Apparently, as discussed on the PR, mo8it has been maintaining
rustlings solo for about nine months, so it makes sense to make him
a lead of the team.

We can handle any other membership adjustments we might want to make
separately and later.