Add repo for Rust-specific soft fork of Enzyme
Currently our submodule for Enzyme in `rust-lang/rust` points to an
external repository outside of our organization.  As we do for LLVM,
it would be useful for us to maintain a Rust-specific soft fork of
Enzyme within `rust-lang/`.  Let's create a repository for this.

In particular, this would unblock some work being held up by CI issues
on the upstream project, and it'd allow us to make some changes that
might better facilitate upstreaming of this work into LLVM.

Probably we should also create a `wg-enzyme` or whatnot team that
could maintain this.  Alternatively or in addition, we could formalize
`wg-autodiff`, for which we have a Zulip stream but no entry here.
But let's handle that separately.  For the moment, we just give access
to the normal teams that would be likely to need to touch it, review
PRs for it, manage issues on it, etc.