Add TC to goals team
We recently created the goals team.  Niko and I had talked, and we had
meant for me to be on this.  Niko had put up an edit on the PR page to
add me.  But apparently this didn't actually make it into a commit,
and then we both missed it on review.  Anyway, this commit adds me to
the goals team.