Move directors to launching pad, give them blog permissions
I think we can trust directors to post on the blog, so lets give them
permissions to do that so they don't need to bug anyone else to do it.

To accommodate the addition of permissions, this promotes it from a
marker team to a team in the launching pad (which we are planning to
rebrand as a home for cross-cutting teams). There's no expectation that
this would change anything in practice, and the membership is still
chosen by the council.