Update lang-docs membership
This updates the lang-docs membership to better reflect the current
state. TC has been helping with the organizational side of lang-docs,
and helping across all of the projects.

I'd like to express a huge appreciation to Gankra, Havvy, and
matthewjasper for their contributions, and they are always welcome to
come back.